tests/wpt/web-platform-tests/url/url-tojson.html

<!doctype html>
<meta charset=utf-8>
<title>URL's toJSON()</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
test(() => {
  const a = new URL("https://example.com/")
  assert_equals(JSON.stringify(a), "\"https://example.com/\"")
})
</script>
