third_party/WebKit/LayoutTests/external/wpt/url/url-tojson-expected.txt

This is a testharness.js-based test.
FAIL URL's toJSON() assert_equals: expected "\"https://example.com/\"" but got "{}"
Harness: the test ran to completion.
